refactor: split utils.py into a utils package
Break the monolithic neural_lam/utils.py (>800 lines, mixing graph
loading, MLP construction, plotting, logging, tensor math and time
helpers) into a neural_lam/utils/ package with one module per concern:

- buffer_list.py  : BufferList
- graph.py        : load_graph + zero-indexing helpers
- networks.py     : make_mlp
- plot.py         : has_working_latex, fractional_plot_bundle
- logging.py      : training logger setup/metrics
- tensor.py       : inverse_softplus, inverse_sigmoid
- time.py         : get_integer_time

The package __init__ re-exports the full public API, so existing
imports (e.g. `from neural_lam import utils`, `from neural_lam.utils
import load_graph`) keep working unchanged. Pure code movement; no
behavioural changes.
